refactor(root): grouped layout components
