+++ src/components/layouts/LayoutExternal/index.tsx
@@ -1,6 +1,6 @@
 import type { ReactNode, FC } from 'react'
 
-import { Header } from '../Header'
+import { Header } from '../../Header'
 
 import { AsideCover } from './parts/AsideCover'
 import { Layout, Main } from './styled'
