+++ src/components/layouts/LayoutExternal/styled.ts
